mx8_install_android: skip MCU demo image install for SMARC builds
Currently M7 is not supported for SMARC.

+++ scripts/sh/var_mk_yocto_sdcard/variscite_scripts/mx8_install_android.sh
@@ -556,9 +556,12 @@ function check_images
 	fi
 
 	rename_remoteproc_images ${soc_name}
-	if [[ ! -f "${imagesdir}/${mcu_os_demo_file}" ]] ; then
-		red_bold_echo "ERROR: ${mcu_os_demo_file} image does not exist"
-		exit 1
+	# mcu demo image is not yet available for VAR-SMARC-MX8M-PLUS SoM
+	if [[ "$soc_name" != *"imx8mp-var-smarc"* ]]; then
+		if [[ ! -f "${imagesdir}/${mcu_os_demo_file}" ]] ; then
+			red_bold_echo "ERROR: ${mcu_os_demo_file} image does not exist"
+			exit 1
+		fi
 	fi
 }
 
@@ -651,10 +654,12 @@ function install_bootloader
 	dd if="${imagesdir}/${bootloader_file}" of="$node" bs=1k seek="${bootloader_offset}" conv=fsync; sync
 
 	echo
-	blue_underlined_bold_echo "Installing mcu demo image: $mcu_os_demo_file"
-
-	dd if="${imagesdir}/${mcu_os_demo_file}" of="${node}" bs=1k seek="${mcu_image_offset}" conv=fsync
-	sync
+	# mcu demo image is not yet available for VAR-SMARC-MX8M-PLUS SoM
+	if [[ "$soc_name" != *"imx8mp-var-smarc"* ]]; then
+		blue_underlined_bold_echo "Installing mcu demo image: $mcu_os_demo_file"
+		dd if="${imagesdir}/${mcu_os_demo_file}" of="${node}" bs=1k seek="${mcu_image_offset}" conv=fsync
+		sync
+	fi
 }
 
 function format_android
